Move models to common folder
